modelado facial/requirements.txt
flask==3.0.3
flask-cors==4.0.1
deepface==0.0.93
numpy==1.26.4
Pillow==10.3.0
opencv-python-headless==4.9.0.80
tf-keras>=2.16.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
